Factor out a secure_store mod for code reuse
