+++ cmd/soroban-cli/src/commands/keys/generate.rs
@@ -9,7 +9,7 @@ use crate::{
     commands::global,
     config::address::KeyName,
     print::Print,
-    signer::keyring::{self, StellarEntry},
+    signer::secure_store::{self, SecureStore},
 };
 
 #[derive(thiserror::Error, Debug)]
@@ -27,7 +27,7 @@ pub enum Error {
     IdentityAlreadyExists(String),
 
     #[error(transparent)]
-    Keyring(#[from] keyring::Error),
+    SecureStore(#[from] secure_store::Error)
 }
 
 #[derive(Debug, clap::Parser, Clone)]
@@ -124,22 +124,8 @@ impl Cmd {
     fn secret(&self, print: &Print) -> Result<Secret, Error> {
         let seed_phrase = self.seed_phrase()?;
         if self.secure_store {
-            // secure_store:org.stellar.cli:<key name>
-            let entry_name_with_prefix = format!(
-                "{}{}-{}",
-                keyring::SECURE_STORE_ENTRY_PREFIX,
-                keyring::SECURE_STORE_ENTRY_SERVICE,
-                self.name
-            );
-
-            //checking that the entry name is valid before writing to the secure store
-            let secret: Secret = entry_name_with_prefix.parse()?;
-
-            if let Secret::SecureStore { entry_name } = &secret {
-                Self::write_to_secure_store(entry_name, seed_phrase, print)?;
-            }
-
-            return Ok(secret);
+            let secret = SecureStore::secret(print, &self.name, seed_phrase)?;
+            return Ok(secret)
         }
         let secret: Secret = seed_phrase.into();
         Ok(if self.as_secret {
@@ -155,24 +141,6 @@ impl Cmd {
         } else {
             secret::seed_phrase_from_seed(self.seed.as_deref())
         }?)
-    }
-
-    fn write_to_secure_store(
-        entry_name: &String,
-        seed_phrase: SeedPhrase,
-        print: &Print,
-    ) -> Result<(), Error> {
-        print.infoln(format!("Writing to secure store: {entry_name}"));
-        let entry = StellarEntry::new(entry_name)?;
-        if let Ok(key) = entry.get_public_key(None) {
-            print.warnln(format!("A key for {entry_name} already exists in your operating system's secure store: {key}"));
-        } else {
-            print.infoln(format!(
-                "Saving a new key to your operating system's secure store: {entry_name}"
-            ));
-            entry.set_seed_phrase(seed_phrase)?;
-        }
-        Ok(())
     }
 }
