+++ cmd/soroban-cli/src/signer.rs
@@ -13,6 +13,7 @@ use crate::xdr::{
 use crate::{config::network::Network, print::Print, utils::transaction_hash};
 
 pub mod keyring;
+pub mod secure_store;
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
